Update WOW64 signed handle bug
NtWow64ReadVirtualMemory64 (and maybe others Native APIs) considers the handles as an unsigned value in Win10. Therefore we can't specify -1 pseudohandle (the current process handle) as a handle value because WOW64 expands it as the unsigned value (uses the mov instruction instead movsxd like in Win7/8). To avoid this restriction it uses the real handle obtained from OpenProcess.